api/requirements-api.txt
fastapi
uvicorn[standard]
python-multipart
ultralytics
pillow
numpy
opencv-python-headless
huggingface_hub
